requirements.server.txt
mcp>=1.2.0
openai>=1.0.0
anthropic>=0.5.0
python-dotenv>=1.0.0
chromadb>=0.4.0